chore: rename types export to artifacts
