+++ src/artifacts/account.ts
@@ -1,0 +1,1 @@
+export type Cid = string
