+++ src/artifacts/index.ts
@@ -2,3 +2,4 @@ export * from './address.js'
 export * from './transaction.js'
 export * from './wallet.js'
 export * from './rpc.js'
+export * from './account.js'
