+++ src/index.ts
@@ -4,4 +4,4 @@ export * from './wallet/index.js'
 export * from './transaction/index.js'
 export * from './token/index.js'
 export * from './account/index.js'
-export * from './types/index.js'
+export * from './artifacts/index.js'
