+++ tests/packages/cjs/src/transaction.js
@@ -1,5 +1,5 @@
 const { Transaction } = require('@zondax/izari-tools/transaction')
-const { Network } = require('@zondax/izari-tools/types')
+const { Network } = require('@zondax/izari-tools/artifacts')
 const assert = require('assert')
 
 async function run() {
